// src/shared/data/regex/vendor/relics/Relic.Gen.ts
export interface RelicRegex {
  id: number
  name: string
  regex: string
  values: number[]
  ranges: number[][]
  affix: 'PREFIX' | 'SUFFIX'
}